pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asyncapi-payload-validator"
version = "1.0.0"
description = "Validate JSON payloads against AsyncAPI specifications with detailed error reporting"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["asyncapi", "validation", "json-schema", "event-driven", "api", "testing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "PyYAML>=6.0",
    "Jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/ingka-group/asyncapi-payload-validator"
Documentation = "https://github.com/ingka-group/asyncapi-payload-validator#readme"
Repository = "https://github.com/ingka-group/asyncapi-payload-validator"
Issues = "https://github.com/ingka-group/asyncapi-payload-validator/issues"

[project.scripts]
asyncapi-validate = "asyncapi_payload_validator.cli:cli"

[tool.setuptools]
packages = ["asyncapi_payload_validator"]

[tool.setuptools.package-data]
asyncapi_payload_validator = ["templates/*.j2"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
